Revamp messages (fix #14)

+++ src/main.rs
@@ -21,10 +21,10 @@ fn main() {
     let database_url = "/tmp/rusty.sqlite3";
     let connection = match Connection::open(database_url) {
         Ok(connection) => connection,
-        Err(e) =>{
-            println!("Error: {} {}", e, database_url);
+        Err(e) => {
+            println!(_log, "Error database connection: {} {}", e, database_url;);
             process::exit(1)
-        },
+        }
     };
     init(&connection); // Initialize database
 
@@ -33,10 +33,13 @@ fn main() {
     let matches = App::from_yaml(yaml).get_matches();
 
     match matches.subcommand() {
-        ("subscribe", Some(sub_matches)) => subscribe(sub_matches, &connection),
+        ("subscribe", Some(sub_matches)) => subscribe(sub_matches, &connection, _log),
         ("unsubscribe", Some(sub_matches)) => unsubscribe(sub_matches, &connection),
-        ("list", Some(_)) => list(&connection),
-        ("update", Some(sub_matches)) => update(sub_matches, &connection),
+        ("list", Some(_)) => list(&connection, _log),
+        ("update", Some(sub_matches)) => {
+            let feed_id = sub_matches.value_of("id").unwrap_or("0").parse::<i64>().unwrap();
+            update(sub_matches, &connection, _log, feed_id);
+        },
         ("pending", Some(_)) => pending(&connection),
         ("download", Some(_)) => download(&connection),
         ("download-dir", Some(sub_matches)) => downloaddir(sub_matches, &connection),
@@ -58,53 +61,61 @@ fn init(connection: &Connection) {
     }
 }
 
-fn subscribe(args: &ArgMatches, connection: &Connection) {
+fn subscribe(args: &ArgMatches, connection: &Connection, _log: slog::Logger) {
     match Url::parse(args.value_of("url").unwrap()) {
         Ok(url) => {
             let subscription = Subscription { id: 0, url: url.as_str().to_string(), label: String::new()};
             connection.execute("insert into subscription (url, label) values (?1, '')", &[&subscription.url]).unwrap();
-            println!("Subscribed to: {}", subscription.url);
-            update(args, connection); 
+            println!("Subscribing to: {}", subscription.url);
+            update(args, connection, _log, connection.last_insert_rowid()); 
+            println!("Subscribed");
         }
         Err(e) => println!("Could not parse url '{}' {}", args.value_of("url").unwrap(), e),
     }
 }
 
 fn unsubscribe(args: &ArgMatches, connection: &Connection) {
+    use std::io;
+    use std::io::prelude::*;
+    let mut stdin = io::stdin();
     let id = args.value_of("id").unwrap();
+    let mut buffer = [0;1];
 
     let mut stmt = connection.prepare("select id, url, label from subscription where id = ?1").unwrap();
     let mut rows = stmt.query_map(&[&id], Subscription::map).unwrap();
     match rows.next() {
         Some(row) => {
             let subscription = row.unwrap();
-            connection.execute("delete from subscription where id = ?1", &[&id]).unwrap();
-            println!("Unsubscribed from: {}", subscription.url);
+            println!("Unsubscribing from: {}", subscription.url);
+            println!("Sure? [y/N]"); 
+            stdin.read_exact(&mut buffer).unwrap();
+            if buffer[0] == 121u8 { // 121 is ascii code for 'y'
+                connection.execute("delete from subscription where id = ?1", &[&id]).unwrap();
+                println!("Unsubscribed from: {}", subscription.url);
+            }
         },
         None => println!("Subscription doesn't exist"),
     }
 }
 
-fn list(connection: &Connection) {
+fn list(connection: &Connection, _log: slog::Logger) {
     let mut stmt = connection.prepare("select id, url, label from subscription").unwrap();
 
+    println!("Subscriptions list:");
     for row in stmt.query_map(&[], Subscription::map).unwrap(){
         let subscription = row.unwrap();
-        println!("- {}: {}", subscription.id, subscription.url);
+        println!("    {}: {} ({})", subscription.id, subscription.label, subscription.url);
     }
 }
 
-fn update(args: &ArgMatches, connection: &Connection) {
+fn update(args: &ArgMatches, connection: &Connection, _log: slog::Logger, feed_id: i64) {
     use hyper::Client; // https://hyper.rs/hyper/v0.10.9/hyper/index.html
     use std::io::Read; // needed for read_to_string trait
     use std::str::FromStr; // needed for FromStr trait on Channel
     use rss::Channel;
 
-    let id = args.value_of("id").unwrap_or("0").parse::<i32>().unwrap();
     let mut stmt = connection.prepare("select id, url, label from subscription").unwrap();
     let client = Client::new(); // create http client
-
-    println!("{:?}", args.is_present("as-downloaded"));
 
     let mut as_downloaded = 0;
     if args.is_present("as-downloaded") {
@@ -113,8 +124,9 @@ fn update(args: &ArgMatches, connection: &Connection) {
 
     for row in stmt.query_map(&[], Subscription::map).unwrap() {
         let subscription = row.unwrap();
+        println!("Updating {}:", subscription.label);
 
-        if id != subscription.id && id > 0 { // if an id is set, update/populate only this id
+        if feed_id != subscription.id && feed_id > 0 { // if an id is set, update/populate only this id
             continue;
         }
 
@@ -123,7 +135,7 @@ fn update(args: &ArgMatches, connection: &Connection) {
         res.read_to_string(&mut body).unwrap(); // extract body from query result
 
         let channel = Channel::from_str(&body).unwrap(); // parse rss into channel
-        connection.execute("update subscription set label = ?1", &[&channel.title()]).unwrap(); // update podcast feed name
+        connection.execute("update subscription set label = ?1 where id = ?2", &[&channel.title(), &subscription.id]).unwrap(); // update podcast feed name
         for item in channel.items() {
             let url = Url::parse(item.enclosure().unwrap().url()).unwrap();
             let  path_segments = url.path_segments().unwrap();
@@ -135,23 +147,28 @@ fn update(args: &ArgMatches, connection: &Connection) {
                     filename = filename + ".mp3";
                     break;
                 }
-                filename = filename + segment;
+                filename = segment.to_string();
             }
 
             let podcast = Podcast { id: 0, subscription_id: subscription.id, url: url.as_str().to_string(), filename: filename};
+            let previous_insert_rowid = connection.last_insert_rowid();
             connection.execute("insert or ignore into podcast (subscription_id, url, filename, downloaded) values (?1, ?2, ?3, ?4)", &[&podcast.subscription_id, &podcast.url, &podcast.filename, &as_downloaded]).unwrap();
-            println!("{:?}", podcast.filename);
+            if previous_insert_rowid != connection.last_insert_rowid() {
+                println!("    New podcast added: {:?}", podcast.filename);
+            }
 
         }
+        println!("Updated");
     }
 }
 
 fn pending(connection: &Connection) {
     let mut stmt = connection.prepare("select id, subscription_id, url, filename from podcast where downloaded = 0").unwrap();
 
+    println!("Pending list:");
     for row in stmt.query_map(&[], Podcast::map).unwrap(){
         let podcast = row.unwrap();
-        println!("- {}: {}", podcast.id, podcast.url);
+        println!("    {} ({})", podcast.filename, podcast.url);
     }
 }
 
@@ -171,11 +188,12 @@ fn download(connection: &Connection) {
     curl.progress(true).unwrap();
     curl.follow_location(true).unwrap();
     curl.progress_function( |a, b, _, _| {
-        print!(" Downloading: {}% \r", (b.mul(100_f64)/a).floor());
+        print!("    Downloading: {}%      \r", (b.mul(100_f64)/a).floor());
         true
     }).unwrap();
     let mut stmt = connection.prepare("select id, subscription_id, url, filename from podcast where downloaded = 0").unwrap();
 
+    println!("Download pending podcast:");
     for row in stmt.query_map(&[], Podcast::map).unwrap(){
         let podcast = row.unwrap();
         let temp = "/tmp/".to_string() + podcast.filename.as_str();
@@ -190,7 +208,7 @@ fn download(connection: &Connection) {
         curl.write_function( move |data| {
             Ok(file.write(data).unwrap())
         }).unwrap();
-        println!("{}", podcast.url);
+        println!("    {}", podcast.filename);
         curl.perform().unwrap();
         connection.execute("update podcast set downloaded = 1, downloaded_at = current_timestamp where id = ?1", &[&podcast.id]).unwrap();
     }
